* Updated PATH in order to find LibVLC dlls in static mediarenderer constructor instead of LibVlcMethods constructor.
* One can now use environment variables in libvlc_searchpath.txt example: %ProgramFiles%\VideaLAN\VLC
