+++ vvvv45/addonpack/src/nodes/plugins/Texture/FileStreamVLC/FileStreamVLC.csproj
@@ -50,7 +50,9 @@
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FileStreamVlcNode.cs" />
+    <Compile Include="FileStreamVlcNode.cs">
+      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
+    </Compile>
     <Compile Include="Player\MediaRenderer.cs" />
     <Compile Include="Player\MemoryToTextureRenderer.cs" />
     <Compile Include="Player\DoubleMemoryBuffer.cs" />
