+++ vvvv45/addonpack/src/nodes/plugins/Texture/FileStreamVLC/FileStreamVlcNode.cs
@@ -47,6 +47,11 @@ using System.Drawing.Drawing2D;
 using System.Text;
 //using Un4seen.Bass;
 //using BassSound.Internals;
+//needed for loading the file with extra search paths
+using System.Windows.Forms;
+using System.Reflection;
+using System.Collections;
+using System.IO;
 
 
 
@@ -233,9 +238,94 @@ namespace VVVV.Nodes.Vlc
 			
 			#endregion MediaRenderer fields
 	
+			#region MediaRenderer static helper functions
+			/// <summary>
+			/// The function returns the file path of the assembly (the dll file) this class resides in. 
+			/// </summary>
+			static public string AssemblyDirectory
+			{
+			    get
+			    {
+			        string codeBase = Assembly.GetExecutingAssembly().CodeBase;
+			        UriBuilder uri = new UriBuilder(codeBase);
+			        string path = Uri.UnescapeDataString(uri.Path);
+			        return Path.GetDirectoryName(path);
+			    }
+			}
+	
+			/// <summary>
+			/// This function will look for the file in all of the folders defined in the searchPathFile, 
+			/// and returns the FIRST path where the given fileName is found. It will also look in the 
+			/// same directory as this dll itself.
+			/// </summary>
+			/// <param name="fileName"></param>
+			/// <param name="searchPathFileName"></param>
+			/// <returns></returns>
+			private static string FindFilePath(string fileName, string searchPathFileName) {
+				//if the libvlc dll is not found in any folder of the PATH environment variable,
+				//search also in directories specified in the text-file
+	
+				string sameDirAsCallingCode = AssemblyDirectory + "\\";
+				if ( File.Exists( sameDirAsCallingCode + fileName) ) {
+					return sameDirAsCallingCode;
+				}
+				
+				//const string searchPathFileName = "libvlc_searchpath.txt";
+				string searchPathFilePath = sameDirAsCallingCode + searchPathFileName;
+				
+				//string searchpath = searchPathFilePath + "\n";
+				try {
+					foreach ( string row in File.ReadAllLines( searchPathFilePath ) ) {
+						//ignore lines starting with # and ignore empty lines
+						if ( ! ( row.Length == 0 || row.StartsWith("#") ) ) {
+							string currentPath = row + ( row.EndsWith( "\\" ) ? "" : "\\" );
+							
+							currentPath = Environment.ExpandEnvironmentVariables( currentPath );
+							
+							if ( row.StartsWith(".") ) {
+								//relative path
+								currentPath = AssemblyDirectory + "\\" + currentPath;
+							}
+							else {
+								//absolute path								
+							}
+	
+							if ( File.Exists( currentPath + fileName) ) {
+								//ideally check if the version is ok to use
+								return currentPath;
+							}
+						}
+					}
+				}
+				catch (IOException) {
+					throw new Exception( "A file named " + searchPathFilePath + " should exist (in the same folder as the Vlc node's dll). This file, which contains paths where the plugin should look for the libvlc.dll (and others) could not be opened, so probably, loading the Vlc plugin will fail." );
+					//MessageBox.Show( "A file named " + searchPathFilePath + " should exist (in the same folder as the Vlc node's dll). This file, which contains paths where the plugin should look for the libvlc.dll (and others) could not be opened, so probably, loading the Vlc plugin will fail.", "Vlc plugin error.", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+				}
+				return null;
+			}
+
+			static void TryToFindLibVLC()
+			{
+				string libvlcdllPath = FindFilePath( "libvlc.dll", "libvlc_searchpath.txt" );
+				if ( libvlcdllPath != null ) {
+					string pathEnvVar = Environment.GetEnvironmentVariable( "PATH" );
+					Environment.SetEnvironmentVariable( "PATH", pathEnvVar + ";" + libvlcdllPath );
+				}
+				else {
+					throw new Exception( "The libvlc.dll file could not be found in any of the paths specified in libvlc_searchpath.txt, so probably, loading the Vlc plugin will fail." );
+					//MessageBox.Show( "The libvlc.dll file could not be found in any of the paths specified in libvlc_searchpath.txt, so probably, loading the Vlc plugin will fail.", "Vlc plugin error.", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+				}
+
+			}
+	
+	        #endregion MediaRenderer static helper functions
+
 			#region MediaRenderer constructor/destructor			
 
 			static MediaRenderer() {
+				
+				TryToFindLibVLC();
+				
 				string[] argv = {
 					"--no-video-title",
 					"--no-one-instance",
@@ -244,6 +334,7 @@ namespace VVVV.Nodes.Vlc
 
 				libVLC = LibVlcMethods.libvlc_new(argv.GetLength(0), argv);				
 			}
+			
 			
 			public MediaRenderer(FileStreamVlcNode parentObject, int index)
 			{
