+++ vvvv45/addonpack/src/nodes/plugins/Texture/FileStreamVLC/libvlc_searchpath.txt
@@ -11,6 +11,6 @@
 # '..\..\VLC' would try to find livlc.dll in c:\VLC and
 # '.\VLC' would try to find livlc.dll in c:\vvvv\plugins\VLC 
 
-C:\Program Files\VideoLAN\VLC
-C:\Program Files (x86)\VideoLAN\VLC
+%ProgramFiles%\VideoLAN\VLC
+%ProgramFilesx86%\VideoLAN\VLC
 .\VLC
